Move batch to its own file
